De-duplicate loading and text helpers shared by TUI and headless
- Extract the local-mode conversation load into loader::load_local, used
  by both the TUI path in main.rs and the headless commands, so they
  apply identical provider/local filtering. Also moves provider_filter
  matching there. (Incidentally clears the single_match lint in main.rs.)
- Promote tool-result text extraction to claude::extract_tool_result_text
  and drop the duplicate copies in the TUI viewer and headless.
- Promote the optional-path-to-string helper to history::path_to_string
  and drop the duplicate copies in conversation_index and headless.

Net -87 lines; behavior unchanged (200 tests pass).
